+++ nautilus_trader/serialization/arrow/serializer.py
@@ -19,6 +19,8 @@ from typing import Any
 
 import pyarrow as pa
 
+from nautilus_trader.common.messages import ComponentStateChanged
+from nautilus_trader.common.messages import TradingStateChanged
 from nautilus_trader.core.correctness import PyCondition
 from nautilus_trader.core.data import Data
 from nautilus_trader.core.message import Event
@@ -38,6 +40,7 @@ from nautilus_trader.persistence.wranglers_v2 import OrderBookDeltaDataWrangler
 from nautilus_trader.persistence.wranglers_v2 import QuoteTickDataWrangler
 from nautilus_trader.persistence.wranglers_v2 import TradeTickDataWrangler
 from nautilus_trader.serialization.arrow.implementations import account_state
+from nautilus_trader.serialization.arrow.implementations import component_events
 from nautilus_trader.serialization.arrow.implementations import instruments
 from nautilus_trader.serialization.arrow.implementations import order_events
 from nautilus_trader.serialization.arrow.implementations import position_events
@@ -292,6 +295,23 @@ register_arrow(
     serializer=order_events.serialize,
     deserializer=order_events.deserialize(OrderFilled),
 )
+
+
+register_arrow(
+    ComponentStateChanged,
+    schema=NAUTILUS_ARROW_SCHEMA[ComponentStateChanged],
+    serializer=component_events.serialize,
+    deserializer=component_events.deserialize(ComponentStateChanged),
+)
+
+
+register_arrow(
+    TradingStateChanged,
+    schema=NAUTILUS_ARROW_SCHEMA[TradingStateChanged],
+    serializer=component_events.serialize,
+    deserializer=component_events.deserialize(TradingStateChanged),
+)
+
 
 for position_cls in PositionEvent.__subclasses__():
     register_arrow(